Namespace for logging tests
