+++ tests/Agent/IntegrationTests/IntegrationTests/Logging/ChangeLogFilenameTests.cs
@@ -7,7 +7,7 @@ using NewRelic.Agent.IntegrationTestHelpers;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests
+namespace NewRelic.Agent.IntegrationTests.Logging
 {
     [NetFrameworkTest]
     public class ChangeLogFilenameTests : IClassFixture<RemoteServiceFixtures.BasicMvcApplicationTestFixture>
